+++ lib/cocoapods/installer/analyzer/pod_variant.rb
@@ -7,6 +7,10 @@ module Pod
         #
         attr_accessor :specs
 
+        # @return [Array<Specification>] the test specs for the target
+        #
+        attr_accessor :test_specs
+
         # @return [Platform] the platform
         #
         attr_accessor :platform
@@ -24,16 +28,21 @@ module Pod
 
         # Initialize a new instance from its attributes.
         #
-        # @param [Array<String>] specs       @see #specs
-        # @param [Platform] platform         @see #platform
-        # @param [Bool] requires_frameworks  @see #requires_frameworks?
+        # @param [Array<Specification>] specs      @see #specs
+        # @param [Array<Specification>] test_specs @see #test_specs
+        # @param [Platform] platform               @see #platform
+        # @param [Bool] requires_frameworks        @see #requires_frameworks?
         #
-        def initialize(specs, platform, requires_frameworks = false)
+        def initialize(specs, test_specs, platform, requires_frameworks = false)
           self.specs = specs
+          self.test_specs = test_specs
           self.platform = platform
           self.requires_frameworks = requires_frameworks
         end
 
+        # @note Test specs are intentionally not included as part of the equality for pod variants since a
+        #       pod variant should not be affected by the number of test specs included.
+        #
         # @return [Bool] whether the {PodVariant} is equal to another taking all
         #         all their attributes into account
         #